Working with character AI. Character can change behaviors according to the current state. E.g. When it stand on flat ground, it can run jump; when it stands next to a rock, it can climb on, push the rock or turn back, etc.

+++ GamePackage/GameObjects/World/World.cpp
@@ -29,7 +29,7 @@ void World::AddVertexObject(int i, int j, Platform * platform)
 
 VertexObject * World::GetVertexObject(int i)
 {
-	if (i >= m_verexObjects.size())
+	if (i >= (int)m_verexObjects.size())
 		return NULL;
 	return m_verexObjects[i];
 }
